requirements.lock.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
asyncmy==0.2.11
attrs==26.1.0
backports.asyncio.runner==1.2.0
banks==2.4.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bidict==0.23.1
build==1.4.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
chromadb==1.5.5
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.13.5
cryptography==46.0.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
docutils==0.22.4
durationpy==0.10
exceptiongroup==1.3.1
fastapi==0.115.9
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.69.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.1
greenlet==3.3.2
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.8.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
ifaddr==0.2.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.13
langchain-core==1.2.23
langchain-google-genai==4.2.1
langchain-openai==1.1.12
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
llama-index-core==0.14.19
llama-index-embeddings-gemini==0.4.2
llama-index-embeddings-openai==0.6.0
llama-index-graph-stores-nebula==0.6.0
llama-index-instrumentation==0.5.0
llama-index-readers-file==0.6.0
llama-index-vector-stores-chroma==0.5.5
llama-index-workflows==2.17.1
Mako==1.3.10
markdown-it-py==4.0.0
markdown2==2.5.5
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nebula3-python==3.8.3
nest-asyncio==1.6.0
networkx==3.4.2
nicegui==2.17.0
nltk==3.9.4
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.30.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.4
playwright==1.58.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pscript==0.7.7
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
PyMuPDF==1.27.2.2
PyMySQL==1.1.2
pyparsing==3.3.2
pypdf==6.9.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.12.1
python-multipart==0.0.22
python-socketio==5.13.0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.3.32
requests==2.33.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.45.3
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
tinytag==2.2.1
tokenizers==0.22.2
tomli==2.4.1
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.34.2
uvloop==0.22.1
vbuild==0.8.2
wait_for2==0.4.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
wsproto==1.3.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
